CreateMonster: Add Vision and Speed panels (moved to Basic Information)
Also includes passive perception in top of basic information panel -
could not fit it in nicely style-wise, with the vision part.

Removed the lone panel or senses and movement - didn't feel natural to
have - and other panels will likely have scroll behaviour anyway, so its
not an issue if the Basic Information panel has it.