View/Services: Add functionality for creating a spell
Makes an external API call (to the Rust backend), this is not
implemented yet - but everything else seems to function as expected.

Unsure if I fully like how I do validation right now, or if I will use a
library in the future, or build a small validation library myself - time
will tell.