API/Monsters: Add a query for getting all monsters
Let us not talk about it okay - I really need to look this over again,
and ensure I didn't add a ton of random shit to everything.

This is also not frontend compatible, and I really need to swap to
data/view representations in the frontend for my own good.

Look into some automatic snake to camel case converts in the frontend as
well - no reason to burden the backend with that dookie.