View/Types: Allow the spellcasting property on Monster to be undefined
This is sort of a hack and there might be a better solution. I need it
be undefined for the backend parsing to function correctly, if none of
the spellcasting properties are set - but unsure if this is the best
approach or not.

The intention is that it is only set to undefined after we have
validated and are on the way to send to backend - so it should be fine.