View/Encounter: Make initial template for creating an encounter
There are still a lot of things that are uncertain:
- The shape of the encounter + entities
- How to handle creation of different entities (although I am thinking a
  modal approach)
- Dropdown menus for setting editing and removing entities

The current intention is that the "details" column should contain
relevant information that is not part of the monster description itself.
Such as the current conditions a monster or player is affected by, or if
they are concentrating, things like that.