View/Components: Place Dialog at fixed position, grow downwards, and make scrollable
Typically, with condition sections in the EncounterBuilder you end up
with a dialog that grows both upwards and downwards because of vertical
centering, and this make it very annoying to work with the UI as the end
user IMO.

This ensures that things like buttons etc only change positions if you
scroll, which is a lot more reasonable.

Might handle max height in a nicer way at some point, but for now hard
coding stuff is easier.