View/Types: Rewrite `TurnReminder` to `CreatureReminder`
The idea here is that instead of having it as part of the combat order
by having these types of reminders be entities by themselves, they are
instead directly linked to other entities.

A monster can have "native" reminders, that are defined as part of the
monster itself, for example a trolls regeneration.

At the same time, the monster entity in combat can have reminders, that
might just be relevant for that specific encounter, or that specific
instance of the monster. These types of reminders also exist on player
entities.

This should in turn make it way easier to deal with combat order in the
future, because the delegate for sorting entities does not have to know
global information to correctly place turn reminders.