crates/views: add `RootView` and `NavigationMenu`
I will more than likely switch out the variant_count crate for strum and
strum_macros in the future. They provide the same functionality, but
strum also provides some enum iterator functionality which is very nice.

There are also some things right now in the views that are fully
placeholder, or will be made more "dynamic" in the future. The
breakcrumbs for example are fully hardcoded and do nothing right now.